requirements.txt
openai>=1.30.0
Pillow>=10.0.0
python-dotenv>=1.0.1
streamlit>=1.44.0
